Update Module to Meet current PSSA guidelines (#29)
There are a number of issues with the current module when run through the standard battery of tests in Appveyor. This PR goes through and fixes these. They are all very minor, mainly formatting and Parameter changes and there shouldn't be any functional impact to the module.

I have also updated the copyright information as per the guidelines and fixed the failing unit tests while bringing them in line with the latest testing template.